app/src/main/java/com/example/androidassignment/domain/usecase/GetTodoUseCase.kt
package com.example.androidassignment.domain.usecase

import com.example.androidassignment.domain.model.Resource
import com.example.androidassignment.domain.model.Todo
import com.example.androidassignment.domain.repository.TodoRepository
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flow
import javax.inject.Inject

class GetTodoUseCase @Inject constructor(
    private val todoRepository: TodoRepository
) {

    operator fun invoke(): Flow<Resource<List<Todo>>> = flow {
        emit(Resource.Loading())
        try {
            todoRepository.getTodos().collect { todos ->
                emit(Resource.Success(todos))
            }
        } catch (e: Exception) {
            emit(Resource.Error(e.message ?: "Unknown error"))
        }
    }
}
